+++ tests/test_utils.py
@@ -6,35 +6,34 @@ from sqlalchemy.sql.sqltypes import Integer
 from asphalt.sqlalchemy.utils import clear_database
 
 
-@pytest.fixture(params=["sqlite_file", "sqlite_memory", "mysql", "postgresql"])
-def engine(request):
-    engine = request.getfixturevalue("%s_engine" % request.param)
-    if engine.dialect.name != "sqlite":
-        engine.execute(CreateSchema("altschema"))
-
-    if request.param != "sqlite-memory":
+@pytest.fixture
+def connection(sync_engine):
+    with sync_engine.connect() as conn:
         metadata = MetaData()
         Table("table", metadata, Column("column1", Integer, primary_key=True))
         Table("table2", metadata, Column("fk_column", ForeignKey("table.column1")))
-        if engine.dialect.name != "sqlite":
+        if conn.dialect.name != "sqlite":
+            conn.execute(CreateSchema("altschema"))
             Table("table3", metadata, Column("fk_column", Integer), schema="altschema")
 
-        metadata.create_all(engine)
+        metadata.create_all(conn)
 
-    yield engine
+        yield conn
 
-    if engine.dialect.name != "sqlite":
-        metadata.drop_all(engine)
-        engine.execute(DropSchema("altschema"))
+        if conn.dialect.name != "sqlite":
+            metadata.drop_all(conn)
+            conn.execute(DropSchema("altschema"))
 
 
-def test_clear_database(engine):
-    clear_database(engine, ["altschema"] if engine.dialect.name != "sqlite" else [])
+def test_clear_database(connection):
+    clear_database(
+        connection, ["altschema"] if connection.dialect.name != "sqlite" else []
+    )
     metadata = MetaData()
-    metadata.reflect(engine)
+    metadata.reflect(connection)
     assert len(metadata.tables) == 0
 
-    if engine.dialect.name != "sqlite":
+    if connection.dialect.name != "sqlite":
         alt_metadata = MetaData(schema="altschema")
-        alt_metadata.reflect(engine)
+        alt_metadata.reflect(connection)
         assert len(alt_metadata.tables) == 0